Support Pass/Fail tasks
Pass/Fail tasks are defined by only defining a fail criteria. These
tasks are never ready, if they would otherwise be ready, they will PASS
instead. This new feature will allow a single pass/fail criteria to be
defined in lieu of defining both a pass and a complemented fail
criteria.

Currently the depends-on tasks have to define a fail criteria which is
the complement of the pass criteria, this is error prone.

Change-Id: If14fe2df89d396a4bd057c10591b3ccc392398df