Support "True" as a task query definition
A query with a "True" definition will always match. This feature allows
users to write tasks that are not meant to be executed.  Tasks with
"True" as fail criteria will always fail and are useful as simple
validation state checks. Tasks with "True" as pass criteria will always
pass when they would otherwise be ready, and are useful for
informational purposes.

Change-Id: I57e490d28b08948e65e31636d6a9b71898b74257