Cache task plugin predicates
Instead of creating task predicates for every change, create them once
and cache them for the lifespan of the attributeFactory. This can have a
significant impact on performance when this lifespan correlates to the
lifespan of the query (Ibc96a754f9c21b21b131d68636274c8231df30cc). This
is particularly valuable when the predicates do expensive bulk queries
up front like

Change-Id: I1616f53dcd44ff96ecf79dfb06ba6ad056b2c343