Separate out building the task definition tree.
The task evaluation and tree building were intertwined. Task definitions
were read as the evaluations were performed and the output tree was
built. Now build the definition tree in a separate TaskTree object which
lazily loads the tree definitions and does not build the tree more than
once. This makes the evaluation code in the TaskAttributeFactory class
much easier to read and follow since this code is now focused on the
evaluation part, the most complex logic to follow.

Change-Id: I4c85b28665c5c7de25492197ea9f51a7fb0da749