Add --task--all flag to the query command
This option causes the query to append task info for all tasks visible
to the current user to the output of each change, whether the tasks are
applicable or not. When this flag is used, an additional 'applicable'
property is included in each task output to indicate whether the task
actually met its applicability criteria or not.

Change-Id: I9cfe4bbf1c46221c63d87af4f7cc309e141f64fa