Update patch set 1
Patch Set 1: Code-Review+1

> Patch Set 1: Verified+1
> 
> ./test/check_task_statuses.sh mfick-lnx 

I'm assuming this needs to be run against a local gerrit instance that has the plugin installed?  Would it be possible to implement the tests as integration tests that can be run with bazel?  Or in a docker environment?

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+1