Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> > Patch Set 9:
> > 
> > (4 comments)
> > 
> > @paladox, is there a different expectation for undefined checks in js? (I really don't ever write js). I typically don't do null checks in Java when the caller can't ever pass a null (even if it would cause a problem for the method).
> > 
> > I guess an approach that would address your concerns would be to move the methods that currently take a 'task' as a parameter to a 'task' object that uses this.<field> instead. Does that work better?
> 
> Polymer 2 needs the undefined checks now.
> 
> We've been fixing this in gerrit's core. But if it works for you using gerrit@master then i guess my comments were unnecessary.


Not always :) The main reason we are adding `undefined` check for polymer 2 is because the changed behavior on observers and computed methods.

```
// polymer 1
computeTest(one, two) {
  // logic will only run one and two both are defined
}

// polymer 2 and above
computeTest(one, two) {
  // logic will run as long as at least one param is defined
}
```

but if you only have one param, it will be unnecessary. And it should fall back to how you treat type in general. And for cases like this, you expect response from the server and you know for sure which field is required / optional, it will be up to you when it is necessary to handle the undefined case. Once we have ts, it will be much easer as we can define interfaces for your API response to make that point very clear, but for compiled js code, it would be exactly the same.


```
interface Task {
  status?: string; // optional, if you use this field, you may need to check for undefined as its optional
  name: string;
  // ...
}

interface PluginInfo {
  name: string;
  roots: Task[];
}

interface TaskResponseItem {
  plugins: PluginInfo[]
}

// then your response will have a type of `TaskResponseItem[]`
```

Then its much more clear whether you need undefined check or not.

Patch-set: 9