Add Bazel support for running docker tests
Now you can do 'bazel test docker-tests' and it will 1) build the task
plugin if needed and 2) run test/docker/run.sh with that built task.jar.

Also fix the test script to exit with the correct status so that bazel
knows if the tests passed or not.

Add Zuul configuration to install Docker on the CI node so we can run
these tests from CI successfully.

Change-Id: I4a06f9fec0de27d90bc70eb54fbcf671305dc0d2