Convert to Polymer 3 using bundled JS artifact
This change de-couples rollup bundle creation from Bazel. rules_nodejs
isn't used, but rollup package is invoked manually.

That way, the developemnt still takes place in seperated JS and HTML
files, but the deployed JS bundle us produced outside of the Bazel
build.

This is done for demonstration purpose only but could also used in real
life. The only disadvantage is that when JavaScript sources are changed
but rollup bundle is not re-created it could be stale.

Development should take place with the following workflow.

1. Change JS sources in gr-task-plugin directory
2. Create JS bundle with npm package manager:

  $ npm install
  $ npm run build

3. Build final plugin artifact:

  $ bazel build task

4. Commit changes to git

Change-Id: Ida1a97ed87cc73ed75696379e81494f2c58c1948