Add ensures to the task tests setups
These ensures verify that certain commands such as json processing don't
error out, and if they do aborts the tests.  This makes developing the
test cases much easier since it becomes more obvious when the test cases
don't parse and it avoids running the tests which are all going to be
broken anyway.

Change-Id: I620debbd23b6d5be11a014ac2deb80c119a1df52