Add a names-factory STATIC Properties test
This testing was previously covered by the names-factory static list
test, however that was not a good place for it since it overloaded that
test with properties tests. The new test covers this properties behavior
and improves upon the coverage significantly. Remove the old coverage
from the names-factory static list, and rename the Namesfactory
Properties test to be inline with the naming of this new test.

Change-Id: I390b66112716d8b10ce18251b3bff1dd696bd3db