Fix to alter change context for TaskChangeFactories
Before this change, the current change was being used for all the task
(queries, properties, etc) evaluations instead of the task's change
returned by the names-factory. Fix this by altering the change context
to take change returned by the names-factory.

Captured performance for "--task--applicable status:open --no-limit"
which remained as it is after this change.

Change-Id: I7430a1d7b016cd0a43043ed2b6eca4395f02c4d5