Fixup some minor task naming/organizational issues.
Fix the missnamed tasksfactory variable which was missing an 's' to be
consistent with the normal task naming.

Fix the addChangeTypeTaskDefinitions() method name which had an extra
's' and was not consistent with the addStaticTypeTaskDefinitions()
method.

Reorder some of the code in the Preloader paths to return early for the
simple cases. This makes the outcome of each condition easier to track.

Change-Id: I09b233e249e4a604bef8aa5447ff8f58cddc005d