Relax looping task property test
Previously this test was strict and would expect a looping property
failure even when the looping properties were unused. Relax this
criteria a bit by ensuring that at least one of the looping properties
under test is used by the task so that it will still trigger property
loop detection even if the property evaluation implementation shortcuts
evaluating unused properties.

Change-Id: Idda7d6a677b6c47345c67c3fa8b853a825454622