Eliminate the Properties.Task inner class
Previously we had a Properties.Task and a Properties.NamesFactory. This
was done mostly so that they could extend the Expander class which was
also defined as an inner class of Properties. However the NamesFactory
class was removed in a previous change because it made calling
conventions a bit clunky when applying the COW pattern. Because they are
tied together and can potentially need to influence each other's
behavior, it also helps to have a single class which handles the
properties for both Tasks and NamesFactories. Now that there is a single
class, it makes more sense to have that class be the top level
Properties class, and not the Properties.Task class. Move all the
Porperties.Task logic to Properties and eliminate Properties.Task.

Change-Id: Ib540986b403c984bed22826e84a08305d7626426