task: Only copy required folders for testing
After the task UI is migrated to polymer, it started
generating node_modules dir which is large and taking
a lot of time to copy into the container during the
initialization of testing framework. So as to overcome
it, only copy required dirs.

Change-Id: I14f58dbf2654211aa727aacf86af6264c0a63e6e
Impact: task plugin docker tests are now faster
Risk-Scope: task plugin tests