Allow tasks to define a duplicate key to identify more duplicates
This allows more taks to be viewed as duplicates at the task
configurators discretion. This can be useful when defining
the key = ${_change_number} to prevent looping on tasks
which preload the same task. This also allows similar "end"
tasks to be defined to behave as if they had looped.

Change-Id: Id9f0311ca3709bd875e2017a4af42938ae83fded