Do not cache TaskTree ChangeNodes
Since these nodes are unlikely to be the used more than once
in a real world situation, do not waste memory caching these
nodes.

Before this change, a task.config which walks all dependencies for a
change would not complete after an hour when run with status:open
--no-limit. At the time of attempting to run the query, it is visible
that lots of CPU is being spent using top, more than a 1000%, likely for
java GC. After this change it completes in around 30mins.

Change-Id: I9fbe4db4b0be9b21a6658892c6e4735371376a76