Explicitly signal end of Task Properties expansion
Previously knowing whether a node's subnodes needed to be reloaded was
only determined at the time a node was refreshed. This was a bit clunky
as it required looking at data from the previous cycle to determine
this. Instead, explicitly signal the completion of node loading so that
this can be evaluated right away without waiting until the next cycle.
This makes the code a bit easier to reason about and less fragile, and
it paves the way for some smarter caching policies which can release
more memory after processing a node instead of the next time the node is
(potentially) loaded.

Change-Id: I875a68f1e988fc75e41a7d6872edec76ef618474