Moving caching logic out of loadSubNodes()
Since loadSubNodes() was no longer following a loader pattern, and it
was getting involved in caching policies which violates normal loader
expectations, move the caching logic back to getSubNodes() and reduce
the complexity of the loaders. Since Node was not previously overriding
getSubNodes(), overriding this now results in a nice place for the
caching logic again and it helps emphasize that it is in fact different
from its superclass' caching policy.

Change-Id: Ibb6dedaf4fc4de4ffb9b763c8bd68c59b66b1974