Use HitHashMap in PredicateCache
Using a common class to gather hit rates makes reporting statistics more
consistent and removes a bunch of code from the PredicateCache class.
This means that the statistic 'misses' will be more appropriately named
'size' now (in simple cases like this, the size always equates to the
misses, so reserve the use of 'misses' for cases when that is not the
true).

Change-Id: I5c21fa9caa4dbaf5f7a28863d24a0070fb4b1320