Use HitHashMap for PreloaderCache
Using a common class to gather hit rates makes reporting statistics more
consistent and removes a bunch of code from the Preloader class. This
means that the statistic 'misses' will be more appropriately named
'size' now (in simple cases like this, the size always equates to the
misses, so reserve the use of 'misses' for cases when that is not the
true).

Change-Id: Ia8b1d8cee6cb96927ab8afe16e6211a7e065c7e5