fixup! Cache Task definition lists for ChangeNodes
The expanded definition was used when creating the definitionList to
cache for ChangeNodes. This resulted in a old, potentially incorrect,
expansion being used when the definitionList was reused. This only
really makes a difference when the subnode happens to use a Change
Property. Fix this bug by storing the unexpanded definitions instead and
add a test case to catch this issue if it happens again.

Change-Id: I38fea30a51e4964fe6993d39a75e55b1d2ee5ad1