Add a TopKeyMap to HitHashMap statistics
A TopKeyMap is a lightweight limited size (5) map with 'long' keys
designed to store only the elements with the top five largest keys. Use
the TopKeyMap in the HitHashMap to track the keys of the elements which
took the longest to load using the StopWatch. Modify all the HitHashMap
users to use the new API and thus benefit from these new statistics.

Change-Id: I19ed563bf6d81f0d72b14906edd676712b8973fe