rename 'untrusted user' to 'non-secret user' for docker tests
Docker tests require a non-admin user that can read task config from
refs/meta/config in the All-Projects project but can not read other
users' config. Before the change, this user is named 'untrusted user'
which is a bit misleading since we trust the user with task config.
Change the user name to 'non-secret user' to be more faithful to its
characteristics.

Change-Id: I25ee1514d9762ce4411a50369b4d153915668ee9