Fix in-tree build failure due to FloggerLogString error
Bazel in-tree build started failing with the error: "Arguments
to log(String) must be compile-time constants or parameters annotated
with @CompileTimeConstant. If possible, use Flogger's formatting log
methods instead", after the introduction of change: I6550cca076 in
stable-3.5 branch. Update log(String) arguments to use Flogger's
formatting, so that the bazel in-tree build is successful.

Change-Id: I6c7fbbec26c82e7ec00bcb533d48d8b82d7b5b88