Provide capability to preview absolute task paths
The '--include-paths' switch allows us to preview the absolute location
of a task. This would be helpful in debugging when we spread out tasks
in multiple repositories in future.

Add .firstTimeRedirect file while setting up the environment for docker
tests to prevent redirection at the very first access to Gerrit
resulting in empty response for the curl call.

Change-Id: I48ef2ac9e214de3e1d7e920b652bfba704445f40