UI: Add chips to ease navigation to the Tasks tab
On changes with a lot of label scores or changed files, users
have to scroll to view status of their tasks. This change adds
a chip (under the commit message, similar to comment chips)
each for 'needs' (aka ready) and 'blocked' (aka failed) tasks
to ease navigation and to show a quick summary of tasks. Here
are some highlights:

 - Chips are shown only after the tasks have loaded
 - Clicking the chips changes focus to the Tasks tab
 - Each chip shows the count of respective tasks
 - Chips are hidden if the count of respective tasks is 0
 - 'needs' chip uses the warning style and 'blocked' uses error

Change-Id: I3e69232661577b64064ce03627cf7efb86728c81