Allow using submit requirement predicates
Update the plugin to allow submit requirement predicates along with
regular change query operators. Organizations using this plugin may
have the need to use submit requirements as tasks to ensure they are
met before the change is ready for CI. One such example is the need
to use submit requirements from code-owners plugin as tasks.

Also, the special 'True' keyword can no longer be used as a task query
definition. A new 'is:true' submit requirement predicate provided by
this plugin should be used instead. This allows us to add test coverage
for submit requirement predicates. When up-merging this change to Gerrit
3.6+, the 'is:true' submit requirement predicate provided by this plugin
should be removed as core[1] implements it.

[1] https://gerrit-review.googlesource.com/c/gerrit/+/323191

Change-Id: Ia17bdd51e5a98b34ae816794a98cb49663d36d61